CSN: CSN-leap-369
- Debug scan service now has a dead link timer that can be set to go off
  at a cycle interval (DEBUG_SCAN_DEADLINK_TIMEOUT).  This is useful
  when the host->FPGA channel deadlocks but the FPGA->host does not.  For
  simplicity, the timer fires whether or not the link is dead.

- Support scratchpad address spaces larger than 32 bits.

+++ modules/leap/libraries/platform-services/mem-services/scratchpad-memory-service/hybrid/hybrid-mem-scratch.h
@@ -97,7 +97,7 @@ class SCRATCHPAD_MEMORY_SERVER_CLASS: public RRR_SERVER_CLASS,
     bool   IsTracing(int level);
 
     // RRR request methods
-    void InitRegion(UINT32 regionID, UINT32 regionEndIdx);
+    void InitRegion(UINT32 regionID, UINT64 regionEndIdx);
 
     void *GetMemPtr(SCRATCHPAD_MEMORY_ADDR addr);
